+++ src/Cookies/SymplifyCookie.php
@@ -17,7 +17,10 @@ final class SymplifyCookie
     private const JSON_COOKIE_NAME = 'sg_cookies';
     private const JSON_COOKIE_VERSION_KEY = '_g';
     private const JSON_COOKIE_VISITOR_ID_KEY = 'visid';
+    private const JSON_COOKIE_PREVIEW_PROJECT_KEY = 'pmr';
+    private const JSON_COOKIE_PREVIEW_VARIATION_KEY = 'pmv';
     private const SUPPORTED_JSON_COOKIE_VERSION = 1;
+
 
     /** @var array<string, mixed> */
     private array $underlying;
@@ -179,6 +182,22 @@ final class SymplifyCookie
         $buf[8] = chr(ord($buf[8]) & 0x3f | 0x80);
 
         return sprintf('%s%s-%s-%s-%s-%s%s%s', ...str_split(bin2hex($buf), 4));
+    }
+
+    public function getPreviewData(): ?array {
+        $projectID = $this->getValue(self::JSON_COOKIE_PREVIEW_PROJECT_KEY);
+
+        if(!is_int($projectID)){
+            return null;
+        }
+
+        $variationID = $this->getValue(self::JSON_COOKIE_PREVIEW_VARIATION_KEY);
+
+        if(!is_int($variationID)){
+            return null;
+        }
+
+        return ['projectID' => $projectID, 'variationID' => $variationID];
     }
 
 }